+++ components/f4jumble/Cargo.toml
@@ -13,3 +13,7 @@ license = "MIT OR Apache-2.0"
 edition = "2018"
 
 [dependencies]
+blake2b_simd = "0.5"
+
+[dev-dependencies]
+proptest = "1"
