+++ components/zcash_address/src/kind/unified.rs
@@ -8,8 +8,6 @@ use zcash_encoding::CompactSize;
 
 use crate::kind;
 
-mod f4jumble;
-
 /// The HRP for a Bech32m-encoded mainnet Unified Address.
 ///
 /// Defined in [ZIP 316][zip-0316].
